+++ changes.md
@@ -12,6 +12,7 @@
 9. Use TGeoManager::fVisLevel in geometry painter
 10. Support "texte" and "texte0" draw options for TH2/TProfile2D classes
 11. Provide radial and z-axis interactive transformation for TGeo drawings
+12. Provide workaround when server returns wrong content length of http ranges request (wrong_http_response)
 
 
 ## Changes in 5.7.1